requirements.txt
click==6.7
python-dotenv==0.16.0
requests==2.20.0
tox==3.23.0